requirements.txt
aiohttp
requests
beautifulsoup4
lxml
fastapi
uvicorn
python-multipart
jinja2

pandas
numpy

sentence-transformers
scikit-learn

google-genai
google-generativeai
python-dotenv
openai
json-repair
pytrends
pillow

tqdm
rich

feedparser
praw

